+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - `Innmind\Http\Header\Custom`
+- `Innmind\Http\Header\Accept\MediaType`
 
 ### Changed
 
@@ -40,6 +41,7 @@
 - `Innmind\Http\Header\Referrer` constructor is now private, use `::of()` named constructor
 - All custom headers now implements `Innmind\Http\Header\Custom`
 - `Innmind\Http\Header` is now a final class
+- `Innmind\Http\Header\Accept` constructor is now private, use `::of()` named constructor
 
 ### Removed
 
@@ -63,6 +65,7 @@
 - `Innmind\Http\Header\LocationValue`
 - `Innmind\Http\Header\RangeValue`
 - `Innmind\Http\Header\ReferrerValue`
+- `Innmind\Http\Header\AcceptValue`
 
 ### Fixed